moved wizard layout into own compoentn
